feat: create a Index page with a template
