+++ gatsby-theme-terminal/index.js
@@ -6,7 +6,5 @@ import Layout, {
   Header,
   PrismCode,
 } from './src/components';
-import IndexPage from './src/pages';
 
 export { Article, componemts, Content, Footer, Header, Layout, PrismCode };
-export default IndexPage;
