+++ gatsby-theme-terminal/src/_templates/posts-template.tsx
@@ -1,8 +1,9 @@
 import { PageProps } from 'gatsby';
-import Layout, { PostsList, PostsData } from '../components';
+import Layout, { PostsList } from '../components';
+import { PostContextData } from '../config/gatsby-node';
 import React from 'react';
 
-const PageTemplate = ({ pageContext }: PageProps<object, PostsData>) => (
+const PageTemplate = ({ pageContext }: PageProps<object, PostContextData>) => (
   <Layout className="Posts-Layout">
     <PostsList keyPrefix="Posts" data={pageContext} />
   </Layout>
