+++ gatsby-theme-terminal/src/components/index.ts
@@ -1,19 +1,10 @@
 import Article from './article';
 import Footer from './footer';
 import Header from './header';
+import PostsList from './posts';
 import PrismCode from './highlight';
 import Layout from './layout';
 import componemts, { Content } from './mdx';
-import PostsList, { PostsData } from './posts';
 
-export {
-  Article,
-  Footer,
-  Header,
-  PostsList,
-  PostsData,
-  PrismCode,
-  componemts,
-  Content,
-};
+export { Article, Footer, Header, PostsList, PrismCode, componemts, Content };
 export default Layout;
